Better parallelism model for sync a Slack channel
