+++ connectors/src/connectors/slack/temporal/workflows.ts
@@ -14,11 +14,9 @@ import { newWebhookSignal } from "./signals";
 const {
   getChannel,
   getChannels,
-  getMessagesForChannel,
   syncThread,
-  syncThreads,
   syncNonThreaded,
-  syncMultipleNoNThreaded,
+  syncChannel,
   getAccessToken,
   fetchUsers,
   saveSuccessSyncActivity,
@@ -70,73 +68,24 @@ export async function syncOneChannel(
   nangoConnectionId: string,
   dataSourceConfig: DataSourceConfig,
   channelId: string,
-  channelName: string,
-  messagesCursor?: string
+  channelName: string
 ) {
   console.log(`Syncing channel ${channelName} (${channelId})`);
-  const threadsToSync: string[] = [];
-  const unthreadedTimeframesToSync = new Map<
-    string,
-    { startTsMs: number; endTsMs: number }
-  >();
 
   const slackAccessToken = await getAccessToken(nangoConnectionId);
+  let messagesCursor: string | undefined = undefined;
 
   do {
-    const messages = await getMessagesForChannel(
-      slackAccessToken,
-      channelId,
-      100,
-      messagesCursor
-    );
-    if (!messages.messages) {
-      // This should never happen because we throw an exception in the activity if we get an error
-      // from the Slack API, but we need to make typescript happy.
-      break;
-    }
-    for (const message of messages.messages) {
-      if (!message.user) {
-        // We do not support messages not posted by users for now
-        continue;
-      }
-      if (message.thread_ts) {
-        if (threadsToSync.indexOf(message.thread_ts) === -1) {
-          // We can end up getting two messages from the same thread if a message from a thread
-          // has also been "posted to channel".
-          threadsToSync.push(message.thread_ts);
-        }
-      } else {
-        const messageTs = parseInt(message.ts as string, 10) * 1000;
-        const weekStartTsMs = getWeekStart(new Date(messageTs)).getTime();
-        const weekEndTsMss = getWeekEnd(new Date(messageTs)).getTime();
-
-        unthreadedTimeframesToSync.set(`${weekStartTsMs}-${weekEndTsMss}`, {
-          startTsMs: weekStartTsMs,
-          endTsMs: weekEndTsMss,
-        });
-      }
-    }
-    await syncThreads(
-      dataSourceConfig,
+    messagesCursor = await syncChannel(
       slackAccessToken,
       channelId,
       channelName,
-      threadsToSync,
-      connectorId
+      dataSourceConfig,
+      connectorId,
+      messagesCursor
     );
-    threadsToSync.length = 0;
-
-    messagesCursor = messages.response_metadata?.next_cursor;
   } while (messagesCursor);
 
-  await syncMultipleNoNThreaded(
-    slackAccessToken,
-    dataSourceConfig,
-    channelId,
-    channelName,
-    Array.from(unthreadedTimeframesToSync.values()),
-    connectorId
-  );
   console.log(`Syncing channel ${channelName} (${channelId}) done`);
 }
